Added softc_entity_t * to get_meta*() functions in collections Added softc_storage_free_storage_strategy() and       softc_storage_strategy_free_datamodel() Fixed some memory problems Cleaned up handling of entities in softpy Updated python tests

+++ foreign/c/test/softc/softc-types-test.cpp
@@ -33,7 +33,7 @@ TEST_F(SoftC_TypesTest, double_array_value)
   const double compare = 1.23;
   softc_double_array_resize_val(da, 10, compare);
   
-  for (int i = 0; i < softc_double_array_size(da); ++i) {
+  for (size_t i = 0; i < softc_double_array_size(da); ++i) {
     ASSERT_EQ(softc_double_array_at(da, i), compare);
   }
   softc_double_array_free(&da);
